*** empty log message ***

+++ tests/paper_eval/builds/setup.sh
@@ -16,7 +16,18 @@ fi
 
 # Build and copy ebwt_build to here
 make -C ${BOWTIE_HOME} ebwt_build
+make -C ${BOWTIE_HOME} ebwt_build_packed
 cp ${BOWTIE_HOME}/ebwt_build .
+cp ${BOWTIE_HOME}/ebwt_build_packed .
+
+if [ ! -f ebwt_build ] ; then
+	echo "Could not build ebwt_build; aborting..."
+	exit 1
+fi
+if [ ! -f ebwt_build_packed ] ; then
+	echo "Could not build ebwt_build_packed; aborting..."
+	exit 1
+fi
 
 # Copy analysis scripts from bowtie dir
 cp ${BOWTIE_HOME}/scripts/summarize_top.pl .
